File classes separation
