+++ NNPTPZ1/NewtonFractals/NewtonFractalsException.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace NNPTPZ1.NewtonFractals
+{
+    class NewtonFractalsException : Exception
+    {
+        public NewtonFractalsException() { }
+
+        public NewtonFractalsException(string message)
+            : base(message)
+        {
+
+        }
+    }
+}
